Reject failed login attempts.
Make sure calls to login() return a rejection when there is an error either connecting to the API or invalid credentials. Without this, even if the login API request fails the promise isn't rejected and so in LoginForm.js the .then() statement is called and makes it look like the user is logged in. With this, the promise is rejected and the .catch() is called.